rename: /interfaces→typs
