+++ components/icons/ShiftIcon.tsx
@@ -1,5 +1,5 @@
 import { COLORS } from "../../constants";
-import { IconProps } from "../../interfaces";
+import { IconProps } from "../../types";
 
 export const ShiftIcon = ({ isChose }: IconProps) => {
   return (
